# Supports serializing a range of rows for CompactRow