… velox/docs/develop/serde/compactrow.rst
… velox/row/benchmark/CMakeLists.txt
… velox/row/benchmark/UnsafeRowSerializeBenchmark.cpp
… velox/row/benchmarks/CMakeLists.txt
… velox/row/benchmarks/DynamicRowVectorDeserializeBenchmark.cpp
… velox/row/benchmarks/UnsafeRowSerializeBenchmark.cpp
… velox/row/tests/CompactRowTest.cpp
… velox/serializers/CompactRowSerializer.cpp
… velox/serializers/benchmarks/CMakeLists.txt
… velox/serializers/benchmarks/RowSerializerBenchmark.cpp
… velox/serializers/tests/CompactRowSerializerTest.cpp