`CompactRow::serialize` only supports serializing the input `RowVector` one row at a time. For workloads that require serializing a range of rows, performance can be improved through vectorized processing. This PR replace the single-row API with a range-based API to handle such cases more efficiently, and it reuse the code from the prev API to handle single-row serialization.
 
Performance improvement for velox_unsafe_row_serialize_benchmark:
```
| Test Name            | Time (us) - before            | Time (us) - after                   | Improvement (%) |
|----------------------|-------------------------------|-------------------------------------|-----------------|
| fixedWidth5          | 48.12                         | 24.50                               | 49.09           |
| fixedWidth10         | 96.62                         | 50.29                               | 47.95           |
| fixedWidth20         | 170.41                        | 86.66                               | 49.15           |
| strings1             | 31.37                         | 24.60                               | 21.58           |
| strings5             | 88.03                         | 72.16                               | 18.03           |
| arrays               | 45.07                         | 36.14                               | 19.81           |
| nestedArrays         | 264.98                        | 257.69                              | 2.75            |
| maps                 | 116.88                        | 111.95                              | 4.22            |
| structs              | 83.24                         | 76.28                               | 8.36            |
```