@marin-ma 2024-08-13
> Thanks. Would you share some benchmark results in the PR description to illustrate the performance improvement?

@rui-mo I have posted the benchmark result into the PR description. The `compact_serialize_xxx` is the original benchmark, and `compact_serialize_range_xxx` is the newly added benchmark with the new API.

```
compact_serialize_fixedWidth5                              48.12us    20.78K
compact_serialize_range_fixedWidth5                        24.50us    40.81K
```