@mbasmanova 2024-08-19
> It seems like the UnsafeRow/CompactRow APIs are specifically designed for presto-on-spark.

@marin-ma UnsafeRow format is Spark-specific and matches Spark. CompactRow was designed specifically for Presto-on-Spark as a more efficient row-wise alternative to UnsafeRow. I'm curious where / how do you use CompactRow? I'd expect Gluten to use UnsafeRow.