@marin-ma 2024-08-20
> I'm curious where / how do you use CompactRow? I'd expect Gluten to use UnsafeRow.

@mbasmanova Gluten uses UnsafeRow for c2r/r2c conversions. Recently, we implemented sort-based shuffle in Gluten. On the shuffle writer side, similar to c2r, it converts the input RowVector into CompactRow, sorts, and then serializes the rows. On the shuffle reader side, it deserializes the data and converts it back to RowVector. Using CompactRow instead of UnsafeRow for intermediate data helps reduce the shuffled data size. We may apply similar optimizations to UnsafeRow in the next step.