@mbasmanova 2024-08-20
@marin-ma Thank you for clarifying.

> converts the input RowVector into CompactRow, sorts, and then serializes the rows.

I wonder why do we need to serialize after sorting? Aren't rows already serialized?

> Using CompactRow instead of UnsafeRow for intermediate data helps reduce the shuffled data size. 

Makes sense. We saw that in PoS as well.