@marin-ma 2024-08-20
> I wonder why do we need to serialize after sorting? Aren't rows already serialized?

@mbasmanova Yes, the rows are serialized. After sorting, we will compress the serialized rows by partition and write it to disk.