@marin-ma 2024-08-22
> but I wonder if we are duplicating code between single- and multi-row serialize methods. Would be nice to consolidate.

@mbasmanova I attempted to consolidate the two methods by extracting and calling the `serializeChildren` from the serialize multiple rows function, but saw significant performance regression for the serialize per row API. Please check this commit https://github.com/marin-ma/velox-oap/commit/d159512710a46da6f8034ffd15d6c0bbd711896e

Benchmark result:
| Test Name            | Time (us) - compact_serialize | Time (us) - compact_serialize (https://github.com/marin-ma/velox-oap/commit/d159512710a46da6f8034ffd15d6c0bbd711896e) |
|----------------------|-------------------------------|-------------------------------------|
| fixedWidth5          | 48.12                         | 214.38                               |
| fixedWidth10         | 96.62                         | 280.59                               |
| fixedWidth20         | 170.41                        | 393.09                               |
| strings1             | 31.37                         | 195.97                               |
| strings5             | 88.03                         | 270.47                               |
| arrays               | 45.07                         | 210.52                               |
| nestedArrays         | 264.98                        | 434.64                              |
| maps                 | 116.88                        | 285.20                              |
| structs              | 83.24                         | 380.90                               |