@mbasmanova 2024-08-22
> I attempted to consolidate the two methods ... but saw significant performance regression for the serialize per row API. 

@marin-ma Would you investigate the source of the regression. Duplicating this logic presents significant maintenance challenges. Let's find a way to avoid copy-paste.