@marin-ma 2024-08-26
@mbasmanova I found the pathlength of the benchmark is about 1.8x larger that before. I tried to debug through adjusting the code, and found that the performance regression is largely due to the additional overhead introduced by the vectorized API. For instance, consider the following snippet of code, there are at least 2 parts introducing extra overheads: 1. the API requires 2 arrays: "nulls" and "offsets". If it's used for only one row, the construction single-element vectors would be non-negligible. 2. Another overhead is from the inner loop `for (auto i = 0; i < rows.size(); ++i) {...}`. The extra loop is useless for one-row situation but it increases the pathlength. 

```
// Serialize a child vector of a row type within a list of rows.
// Write the serialized data at offsets of buffer row by row.
// Update offsets with the actual serialized size.
template <TypeKind kind>
void serializeTyped(
    const raw_vector<vector_size_t>& rows,
    uint32_t childIdx,
    DecodedVector& decoded,
    size_t valueBytes,
    const raw_vector<uint8_t*>& nulls,
    char* buffer,
    std::vector<size_t>& offsets) {
  const auto mayHaveNulls = decoded.mayHaveNulls();
  const auto* rawData = decoded.data<char>();

  for (auto i = 0; i < rows.size(); ++i) {
    if (mayHaveNulls && decoded.isNullAt(rows[i])) {
      bits::setBit(nulls[i], childIdx, true);
    } else {
      // Write fixed-width value.
      memcpy(
          buffer + offsets[i],
          rawData + decoded.index(rows[i]) * valueBytes,
          valueBytes);
    }
    offsets[i] += valueBytes;
  }
}
```

Despite my efforts, I couldn't find a better way to unify the implementation without causing the performance regression. Could you suggest a more efficient approach to achieve this? Thank you!