@mbasmanova 2024-08-28
@marin-ma BTW, do we still need single-row API? If not, perhaps, removing that API altogether is a quick way to unblock.